Migrate enzyme PasswordConfirmation tests to React Testing Library
